cloud-functions/functions/get_area/requirements.txt
google-cloud-bigquery==3.0.1

Fiona==1.8.21
Shapely==1.8.1.post1
